fix diagonal tilt bug and subtle frame stutter
previously, the game would prioritize x movement over y movement, even if you were tilted in the y direction more.  this is fixed.

also, the app no longer waits a frame each time it moves.  before this was fixed, this would mean that, assuming the corridorWidth was 8, it would roll down a long hallway spending 9 frames per unit movement, moving 1 px 8 frames of the time, but halting and moving 0px for a single frame.