actually fix movement, fix select speedbug
funnily enough, pressing the select button again and again would start more gameloops.  you could play maze at hyperspeed.  now movement actually prioritizes the direction you're leaning, but not to the exclusion of other directions.  it used to be that if there was no wall to the right or north of you and you were leaning northnorthwest, it would still go right instead.  after the first attempted fix, it would go north in that situation, but if there was a wall to the north, it actually would refuse to go west, which is even worse.  now this fixes everything (as far as I know)